requirements.txt
flask
flask-cors
torch
diffusers
transformers
Pillow
gunicorn

